chore(api): move validation-treatments.js to evaluation scope
