+++ api/lib/domain/services/user-reconciliation-service.js
@@ -5,6 +5,10 @@ const { pipe } = fp;
 import randomString from 'randomstring';
 
 import { isOneStringCloseEnoughFromMultipleStrings } from '../../../src/evaluation/domain/services/string-comparison-service.js';
+import {
+  normalizeAndRemoveAccents,
+  removeSpecialCharacters,
+} from '../../../src/evaluation/domain/services/validation-treatments.js';
 import { LEVENSHTEIN_DISTANCE_MAX_RATE, STUDENT_RECONCILIATION_ERRORS } from '../../../src/shared/domain/constants.js';
 import {
   AlreadyRegisteredUsernameError,
@@ -13,7 +17,6 @@ import {
   OrganizationLearnerAlreadyLinkedToUserError,
 } from '../../../src/shared/domain/errors.js';
 import { areTwoStringsCloseEnough } from '../../../src/shared/domain/services/string-comparison-service.js';
-import { normalizeAndRemoveAccents, removeSpecialCharacters } from './validation-treatments.js';
 
 const STRICT_MATCH_RATIO = 0;
