+++ api/tests/evaluation/unit/domain/services/validation-treatments_test.js
@@ -3,8 +3,8 @@ import {
   applyTreatments,
   normalizeAndRemoveAccents,
   removeSpecialCharacters,
-} from '../../../../lib/domain/services/validation-treatments.js';
-import { expect } from '../../../test-helper.js';
+} from '../../../../../src/evaluation/domain/services/validation-treatments.js';
+import { expect } from '../../../../test-helper.js';
 
 describe('Unit | Service | Validation Treatments', function () {
   describe('#normalizeAndRemoveAccents', function () {
